Backend/requirements.txt
fastapi[all]
pymongo
motor
pydantic[dotenv]
passlib
bcrypt
python-jose